+++ planning/autoware_velocity_smoother/src/resample.cpp
@@ -73,7 +73,7 @@ TrajectoryPoints resampleTrajectory(
   bool is_endpoint_included = false;
   for (size_t i = 1; static_cast<double>(i) <= N; ++i) {
     double ds = ds_nominal;
-    if (i > Nt) {
+    if (static_cast<double>(i) > Nt) {
       // if the planning time is not enough to see the desired distance,
       // change the interval distance to see far.
       ds = std::max(param.sparse_min_interval_distance, param.sparse_resample_dt * v_current);
@@ -93,7 +93,7 @@ TrajectoryPoints resampleTrajectory(
     }
 
     // Check if the distance is longer than minimum_trajectory_length
-    if (i > Nt && dist_i >= param.min_trajectory_length) {
+    if (static_cast<double>(i) > Nt && dist_i >= param.min_trajectory_length) {
       if (
         std::fabs(out_arclength.back() - (param.min_trajectory_length + front_arclength_value)) <
         1e-3) {
