+++ planning/autoware_velocity_smoother/src/smoother/l2_pseudo_jerk_smoother.cpp
@@ -138,7 +138,7 @@ bool L2PseudoJerkSmoother::apply(
   b is almost 0, and is not a big problem.
   */
   for (unsigned int i = 0; i < N; ++i) {
-    const int j = 2 * N + i;
+    const int j = static_cast<int>(2 * N + i);
     A(i, i) = 1.0;   // b_i
     A(i, j) = -1.0;  // -delta_i
     upper_bound[i] = v_max[i] * v_max[i];
@@ -147,7 +147,7 @@ bool L2PseudoJerkSmoother::apply(
 
   // a_min < a - sigma < a_max
   for (unsigned int i = N; i < 2 * N; ++i) {
-    const int j = 2 * N + i;
+    const int j = static_cast<int>(2 * N + i);
     A(i, i) = 1.0;   // a_i
     A(i, j) = -1.0;  // -sigma_i
     if (i != N && v_max[i - N] < std::numeric_limits<double>::epsilon()) {
@@ -185,7 +185,8 @@ bool L2PseudoJerkSmoother::apply(
 
   const auto tf1 = std::chrono::system_clock::now();
   const double dt_ms1 =
-    std::chrono::duration_cast<std::chrono::nanoseconds>(tf1 - ts).count() * 1.0e-6;
+    static_cast<double>(std::chrono::duration_cast<std::chrono::nanoseconds>(tf1 - ts).count()) *
+    1.0e-6;
 
   // execute optimization
   const auto ts2 = std::chrono::system_clock::now();
@@ -210,8 +211,8 @@ bool L2PseudoJerkSmoother::apply(
     double v = optval.at(i);
     // std::cout << "[smoother] v[" << i << "] : " << std::sqrt(std::max(v, 0.0)) <<
     // ", v_max[" << i << "] : " << v_max[i] << std::endl;
-    output.at(i).longitudinal_velocity_mps = std::sqrt(std::max(v, 0.0));
-    output.at(i).acceleration_mps2 = optval.at(i + N);
+    output.at(i).longitudinal_velocity_mps = static_cast<float>(std::sqrt(std::max(v, 0.0)));
+    output.at(i).acceleration_mps2 = static_cast<float>(optval.at(i + N));
   }
   for (unsigned int i = N; i < output.size(); ++i) {
     output.at(i).longitudinal_velocity_mps = 0.0;
@@ -231,7 +232,8 @@ bool L2PseudoJerkSmoother::apply(
 
   const auto tf2 = std::chrono::system_clock::now();
   const double dt_ms2 =
-    std::chrono::duration_cast<std::chrono::nanoseconds>(tf2 - ts2).count() * 1.0e-6;
+    static_cast<double>(std::chrono::duration_cast<std::chrono::nanoseconds>(tf2 - ts2).count()) *
+    1.0e-6;
   RCLCPP_DEBUG(logger_, "init time = %f [ms], optimization time = %f [ms]", dt_ms1, dt_ms2);
 
   return true;
